В основном пострадал WatchService
• Косметические изменения.
• Добавлен HELP.md
• WatchServise сделана синглтоном.
• Изменена синхронизация контроллера со службой наблюдения: теперь для этого используется не Lock, а приостановка службы на время выполнения операции, ход которой не должен отслеживаться службой.
• Улучшена обработка событий внутри потока службы наблюдения.
• Лоакльный Манипулятор не удаляется при отключении.

+++ client/src/main/java/ru/gb/simplenas/client/services/NetClient.java
@@ -24,7 +24,7 @@ public interface NetClient extends Runnable {
 
     FileInfo fileInfo (String folder, String fileName);
 
-    int countFolderEntries (String strParent, FileInfo fi);
+    long countFolderEntries (String strParent, FileInfo fi);
 
     NasMsg delete (String strParent, FileInfo fi);
 }
